Support building JS, publishing to Github Packages, and installing globally
This is probably good to go now. Need to add some short instructions for how to install globally.

Basically you need to get a Github token for installing packages, then setup your global .npmrc file. ~/.npmrc looks like this:

//npm.pkg.github.com/:_authToken=${GITHUB_PACKAGE_TOKEN}
@liveramp:registry=https://npm.pkg.github.com/

Then you can install reslang globally with

yarn global add @liveramp/reslang

Needed to change the name to @liveramp/reslang because Github Packages requires the package to be prefixed with the org name. If this is a dealbreaker, we can figure something else out.