Nullable pagination strings (#174)
* Make pagination fields nullable

According to RFC API-3, these pagination fields are nullable strings. When the type is only specified as a string, then the possible values are string | undefined.

In OpenAPI 3.1, they will use normal JSON schema (array of [“string”, “null”], for example). But in 3.0.x they still have their own flavor of it with a `nullable` field.